Set up based API
The api folder is special as it needs to be called api. It will be treated like a normal page so you can navigate to it but you would not use it like that. Instead it would be used to make requests evident by the code. The function isnt a normal react rafce function but an api function which has req and res in the params. To navigate to it you go to /api/feedback